fixed progress bar visibility
